requirements.txt
alembic==1.16.2
bcrypt==4.3.0
bidict==0.23.1
blinker==1.9.0
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
Flask==3.1.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
fonttools==4.58.4
greenlet==3.2.3
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
numpy==2.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.2
qrcode==8.2
reportlab==4.4.2
scikit-learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.41
threadpoolctl==3.6.0
typing_extensions==4.14.0
tzdata==2025.2
Werkzeug==3.1.3
wsproto==1.2.0
xlsxwriter==3.2.5
